+++ app/Http/Controllers/SiteController.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Http\Request;
 use App\Models\Site;
 use App\Models\Hoster;
 use App\Models\User;
@@ -19,16 +19,47 @@ class SiteController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
         // Sites list
+        $campaign_id = $request->query('campaign_id');
+        $hoster_id = $request->query('hoster_id');
+        $hoster_id_domain = $request->query('hoster_id_domain');
+        $user_id = $request->query('user_id');
+        $search_domain = $request->query('search_domain');
+
+        $view = 'sites.list';
+
+        $sites = Site::orderBy('id', 'desc');
+
         if (Auth::user()->role == 1) {
-            $sites = Site::paginate(50);
-            return view('sites.list', compact('sites'));
+            if (!empty($user_id)) {
+                $sites = $sites->where('user_id', $user_id);
+            }
         } else {
-            $sites = Site::paginate(50);
-            return view('sites.list', compact('sites'));
+            $sites = $sites->where('user_id', Auth::user()->id);
         }
+        
+
+        if (!empty($search_domain)) {
+            $sites = $sites->where('domain', 'LIKE', '%'.$search_domain.'%');
+        }
+
+        if (!empty($campaign_id)) {
+            $sites = $sites->where('campaign_id', $campaign_id);
+        }
+
+        if (!empty($hoster_id)) {
+            $sites = $sites->where('hoster_id', $hoster_id);
+        }
+
+        if (!empty($hoster_id_domain)) {
+            $sites = $sites->where('hoster_id_domain', $hoster_id_domain);
+        }
+
+        $sites = $sites->paginate(100);
+
+        return view($view, compact('sites', 'search_domain'));
     }
 
     /**
@@ -39,11 +70,14 @@ class SiteController extends Controller
     public function create()
     {
         // Adding new site page
-        $users = User::all();
-        $hosters = Hoster::all();
-        $campaigns = Campaign::all();
+        if (Auth::user()->role == 1) {
+            $users = User::all();
+            $hosters = Hoster::all();
+            $campaigns = Campaign::all();
 
-        return view('sites.create', compact('users', 'hosters', 'campaigns'));
+            return view('sites.create', compact('users', 'hosters', 'campaigns'));
+        }
+        return redirect()->route('sites.list');
     }
 
     /**
@@ -55,33 +89,36 @@ class SiteController extends Controller
     public function store(Request $request)
     {
         //
-        $data = $this->validate($request, [
-            'domain' => 'required|unique:sites',
-            'user_id' => 'numeric',
-            'campaign_id' => 'required|numeric',
-            'hoster_id' => 'required|numeric',
-            'hoster_id_domain' => 'required|numeric',
-            'ftp_host' => 'nullable',
-            'ftp_user' => 'nullable',
-            'ftp_pass' => 'nullable',
-            'yandex' => 'nullable|numeric',
-            'facebook' => 'nullable|numeric'
-        ]);
+        if (Auth::user()->role == 1) {
+            $data = $this->validate($request, [
+                'domain' => 'required|unique:sites',
+                'user_id' => 'numeric',
+                'campaign_id' => 'required|numeric',
+                'hoster_id' => 'required|numeric',
+                'hoster_id_domain' => 'required|numeric',
+                'ftp_host' => 'nullable',
+                'ftp_user' => 'nullable',
+                'ftp_pass' => 'nullable',
+                'yandex' => 'nullable|numeric',
+                'facebook' => 'nullable|numeric'
+            ]);
 
-        $site = new Site();
-        $site->fill($data);
-        $title = $site->domain;
+            $site = new Site();
+            $site->fill($data);
+            $title = $site->domain;
 
-        $site->domain = str_replace('http://', '', $site->domain);
-        $site->domain = str_replace('https://', '', $site->domain);
+            $site->domain = str_replace('http://', '', $site->domain);
+            $site->domain = str_replace('https://', '', $site->domain);
 
-        if ($request->clean_host == 1) {
-            $this->cleanHost($site);
+            if ($request->clean_host == 1) {
+                $this->cleanHost($site);
+            }
+
+            $site->save();
+
+            return redirect()->route('sites.list')->with('message', "Сайт с адресом «" . $title . "» был добавлен в таблицу.");
         }
-
-        $site->save();
-
-        return redirect()->route('sites.list')->with('message', "Сайт с адресом «" . $title . "» был добавлен в таблицу.");
+        return redirect()->route('sites.list');
     }
 
     /**
@@ -104,12 +141,15 @@ class SiteController extends Controller
     public function edit($id)
     {
         //
-        $site = Site::findOrFail($id);
-        $users = User::all();
-        $hosters = Hoster::all();
-        $campaigns = Campaign::all();
+        if (Auth::user()->role == 1) {
+            $site = Site::findOrFail($id);
+            $users = User::all();
+            $hosters = Hoster::all();
+            $campaigns = Campaign::all();
 
-        return view('sites.edit', compact('site', 'users', 'campaigns', 'hosters'));
+            return view('sites.edit', compact('site', 'users', 'campaigns', 'hosters'));
+        }
+        return redirect()->route('sites.list');
     }
 
     /**
@@ -122,29 +162,32 @@ class SiteController extends Controller
     public function update(Request $request, $id)
     {
         //
-        $site = Site::findOrFail($id);        
-        $data = $this->validate($request, [
-            'domain' => 'required|unique:sites,domain,' . $site->id,
-            'user_id' => 'numeric',
-            'campaign_id' => 'required|numeric',
-            'hoster_id' => 'required|numeric',
-            'hoster_id_domain' => 'required|numeric',
-            'ftp_host' => 'nullable',
-            'ftp_user' => 'nullable',
-            'ftp_pass' => 'nullable',
-            'yandex' => 'nullable',
-            'facebook' => 'nullable'
-        ]);
+        if (Auth::user()->role == 1) {
+            $site = Site::findOrFail($id);        
+            $data = $this->validate($request, [
+                'domain' => 'required|unique:sites,domain,' . $site->id,
+                'user_id' => 'numeric',
+                'campaign_id' => 'required|numeric',
+                'hoster_id' => 'required|numeric',
+                'hoster_id_domain' => 'required|numeric',
+                'ftp_host' => 'nullable',
+                'ftp_user' => 'nullable',
+                'ftp_pass' => 'nullable',
+                'yandex' => 'nullable',
+                'facebook' => 'nullable'
+            ]);
 
-        $site->fill($data);
+            $site->fill($data);
 
-        $site->domain = str_replace('https://', '', $site->domain);
-        $site->domain = str_replace('http://', '', $site->domain);
-        $site->save();
+            $site->domain = str_replace('https://', '', $site->domain);
+            $site->domain = str_replace('http://', '', $site->domain);
+            $site->save();
 
-        $updateSettings = $this->updateSettingsAfterUpdateSite($site);
-        
-        return redirect()->route('sites.list')->with('message', "Сайт «" . $site->domain . "» был обновлён.");;
+            $updateSettings = $this->updateSettingsAfterUpdateSite($site);
+            
+            return redirect()->route('sites.list')->with('message', "Сайт «" . $site->domain . "» был обновлён.");
+        }
+        return redirect()->route('sites.list');
     }
 
     /**
@@ -156,13 +199,14 @@ class SiteController extends Controller
     public function destroy($id)
     {
         //
-        $site = Site::findOrFail($id);
-        if ($site) {
-            $title = $site->domain;
-            $site->delete();
-            return redirect()->route('sites.list')->with('message', "Сайт «" . $title . "» был удалён из базы.");
+        if (Auth::user()->role == 1) {
+            $site = Site::findOrFail($id);
+            if ($site) {
+                $title = $site->domain;
+                $site->delete();
+                return redirect()->route('sites.list')->with('message', "Сайт «" . $title . "» был удалён из базы.");
+            }
         }
-
         return redirect()->route('sites.list');
     }
 
@@ -229,5 +273,13 @@ class SiteController extends Controller
         
         $ftp = new Flysystem($site);
         $save = $ftp->saveSettingsJson($json);
+    }
+
+    public function transfer(Request $request)
+    {
+        $site = Site::findOrFail($request->siteid);
+        $site->user_id = $request->user_id;
+        $site->save();
+        return redirect()->route('sites.list')->with('message', "Сайт $site->domain успешно передан другому пользователю");
     }
 }
